Inline AlcesStatic in Alces namespace
This is the only place AlcesStatic is used, so there seems no need to
have this exist as a separate module - this just adds some indirection
and makes it slightly harder to find methods methods. Inlining it in
Alces does make the class a bit larger, but the class was already large
and doing a lot, and this just makes this more explicit.
