+++ lib/underware/namespaces/alces.rb
@@ -1,6 +1,7 @@
 
 # frozen_string_literal: true
 
+require 'hashie'
 require 'ostruct'
 
 require 'underware/exceptions'
@@ -19,11 +20,13 @@ Underware::Utils::DynamicRequire.relative('.')
 module Underware
   module Namespaces
     class Alces
-      include Mixins::AlcesStatic
-
       NODE_ERROR = 'Error, a Node is not in scope'
       GROUP_ERROR = 'Error, a Group is not in scope'
       DOUBLE_SCOPE_ERROR = 'A node and group can not both be in scope'
+      LOCAL_ERROR = <<-EOF.strip_heredoc
+        The local node has not been configured Please run: `underware
+        configure local`
+      EOF
 
       delegate :config, :answer, to: :scope
       attr_reader :platform
@@ -49,6 +52,66 @@ module Underware
       def initialize(platform: nil)
         @platform = platform&.to_sym
         @stacks_hash = {}
+      end
+
+      def alces
+        self
+      end
+
+      def domain
+        @domain ||= Namespaces::Domain.new(alces)
+      end
+
+      def nodes
+        @nodes ||= begin
+          arr = NodeattrInterface.all_nodes.map do |node_name|
+            Namespaces::Node.new(alces, node_name)
+          end
+          Namespaces::UnderwareArray.new(arr)
+        end
+      end
+
+      def groups
+        @groups ||= begin
+          arr = group_cache.map do |group_name|
+            index = group_cache.index(group_name)
+            Namespaces::Group.new(alces, group_name, index: index)
+          end
+          Namespaces::UnderwareArray.new(arr)
+        end
+      end
+
+      def data
+        DataFileNamespace.new
+      end
+
+      def local
+        @local ||= begin
+          unless nodes.respond_to?(:local)
+            raise UninitializedLocalNode, LOCAL_ERROR
+          end
+          nodes.local
+        end
+      end
+
+      def build_interface
+        @build_interface ||= determine_build_interface
+      end
+
+      def orphan_list
+        @orphan_list ||= group_cache.orphans
+      end
+
+      def questions
+        @questions ||= loader.question_tree
+      end
+
+      def assets
+        @assets ||= AssetArray.new(self)
+      end
+
+      def asset_cache
+        @asset_cache ||= Underware::Cache::Asset.new
       end
 
       def node
@@ -149,6 +212,50 @@ module Underware
 
       def wrapped_binding
         Templating::NilDetectionWrapper.wrap(self)
+      end
+
+      def group_cache
+        @group_cache ||= GroupCache.new
+      end
+
+      def loader
+        @loader ||= Validation::Loader.new
+      end
+
+      def determine_build_interface
+        if config.configured_build_interface&.present?
+          config.configured_build_interface
+        else
+          # Default to first network interface if `build_interface` has not
+          # been configured by user.
+          Network.interfaces.first
+        end
+      end
+
+      class DataFileNamespace
+        delegate :namespace_data_file, to: FilePath
+
+        def method_missing(message, *_args)
+          data_file_path = namespace_data_file(message)
+          if respond_to?(message)
+            Hashie::Mash.load(data_file_path)
+          else
+            # Normally `method_missing` should call `super` if it doesn't
+            # `respond_to?` a message. In this case this is a namespace
+            # designed to be used by users writing templates, so give them an
+            # informative error message for what they've probably missed
+            # instead. This does mean though that we could get a confusing
+            # error message if something else goes wrong in this class, so I
+            # could eventually come to regret this.
+            raise UserUnderwareError,
+                  "Requested data file doesn't exist: #{data_file_path}"
+          end
+        end
+
+        def respond_to_missing?(message, _include_all = false)
+          data_file_path = namespace_data_file(message)
+          File.exist?(data_file_path)
+        end
       end
     end
   end
